The simulation with a first policy seems to be running! I think I was able to replicate the example problem with two part types and different efficiencies, although now I need to measure e1 and e2.
Corrected a bug in which the Event handler was ignoring that the machine could be down when computing the cumulative production. Also, Schedule objects now have a dumpable property that can be set during construction (default is false). If a schedule is dumpable, we can erase all programmed events. This is useful for removing events from the production schedule after a failure, so that we don't get an agglomeration of ControlEvent's.